adds basic authorization

+++ app/resources/views/shared/show.blade.php
@@ -2,15 +2,20 @@
 @section('content')
 
   @section('panel_body')
+
     {{ $resource }}
     {{-- {!! HTML::image($resource->imagePath() . $resource->image, $resource->title, ['class' => 'img-rounded img-responsive img-thumbnail']) !!} --}}
   @endsection {{-- end panel_body --}}
 
   @section('panel_actions')
   <div class="pull-right">
-      {!! offer_link($resource) !!}
-      {!! edit_link($resource) !!}
-      {!! destroy_link($resource) !!}
+      @if(Auth::user()->canOffer($resource))
+        {!! offer_link($resource) !!}
+      @endif
+      @if(Auth::user()->canEdit($resource))
+        {!! edit_link($resource) !!}
+        {!! destroy_link($resource) !!}
+      @endif
   </div>
   @endsection {{-- end panel_actions --}}
